User registration and signin completed

+++ app/Http/Controllers/MeetingController.php
@@ -5,9 +5,18 @@ namespace App\Http\Controllers;
 use App\Meeting;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\JWTAuth;
+//use JWTAuth;
 
 class MeetingController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('jwt.auth', ['only'=>[
+            'update', 'store', 'destroy'
+        ]]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -15,11 +24,10 @@ class MeetingController extends Controller
      */
     public function index()
     {
-        $meeting = Meeting::all();
+        $meetings = Meeting::all();
 
         foreach ($meetings as $meeting) {
             $meetings->view_meeting = [
-                //'href' => 'api/v1/meeting/'.$meeting->id,
                 'href' => 'api/v1/meeting/1',
                 'method' => 'GET'
             ];
@@ -27,8 +35,8 @@ class MeetingController extends Controller
         $response = [
             'msg' => '>List of all meetings',
             'meetings' => [
-                $meeting,
-                $meeting
+                $meetings,
+                $meetings
             ]
         ];
 
@@ -46,14 +54,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -107,29 +118,20 @@ class MeetingController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //return 'It works fine!';
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
-
-        /*$meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];*/
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -165,6 +167,15 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+
+        if($meeting->users()->where('users.id', $user->id)->first()){
+            return response()->json(['msg' => 'User not registered for the meeting, update not successful!'], 401);
+        }
+
         $users = $meeting->users;
         $meeting->users()->detach();
